Projects

OSM projects are projects that use tags in some way, e.g. editors, routers, maps, or data extractions tools. Taginfo keeps information about the keys and tags that a project uses. These functions provide a means to retrieve this information.

projects_all(query = NULL, status = "OK", ...)

projects_keys(query = NULL, ...)

projects_tags(query = NULL, ...)

project_icon(project, path = tempfile(), plot = TRUE)

project_tags(project, ...)

Source

https://taginfo.openstreetmap.org/taginfo/apidoc

Arguments

query

A search term to filter results by. For projects_all, this matches name or description. For projects_keys, this matches the keys. For projects_tags, this matches the tags.

status

Only show projects with a given status. This argument does not seem to do anything currently.

...

Additional generic parameters used for pagination, filtering, and sorting. Can include the following:

page: If pagination is enabled, specifies the page number.

rp: Enables pagination and specifies the number of results returned per page.

filter: A character vector specifying the OSM geometry type to filter the results by. For example, if "nodes", only tags on nodes are returned. Can be several of "all" (no filtering), "nodes", "relations", or "ways".

sortname: Numeric column of the output by which to arrange the results.

sortorder: Whether to sort ascending ("asc") or descending ("desc").

Note that not all additional arguments are defined for all API methods. For details, refer to api_spec.

project

A project ID.

path

Path to a file where the plot should be written to.

plot

Whether to plot the distribution using the R graphics device using show_image. If FALSE, simply returns the path to the file. Defaults to TRUE.

Value

projects_all: A dataframe containing basic information about all projects and their used keys

projects_key / projects_tags: A dataframe containing information about each key / tag associated with a project and their occurences in projects and wiki pages.

project_icon: The path to the downloaded icon and (if plot = TRUE), the icon shown as an R plot. Note that some icons are stored as SVG and produce an error when trying to plot.

project_tags: A dataframe containing information about how a tag is used in a project.

Examples

# get all projects in osmtags
projects_all()
#> # A tibble: 230 × 10
#>    id     name  project_url icon_url doc_url description key_entries tag_entries
#>    <chr>  <chr> <chr>       <chr>    <chr>   <chr>             <int>       <int>
#>  1 archi… Arch… https://os… https:/… https:… Data about…          11           0
#>  2 babyk… Baby… https://ba… https:/… https:… A map for …           2          30
#>  3 bano   BANO… http://ope… NA       NA      Open Adres…           4           0
#>  4 spiel… Berl… https://os… https:/… https:… Interactiv…          49          13
#>  5 bexhi… Bexh… https://be… https:/… https:… A highly d…          38           1
#>  6 bikec… Bike… https://ma… https:/… NA      Bicycle ro…          26          96
#>  7 boxlo… Box … https://bo… https:/… NA      Find vendi…           1          46
#>  8 bubat… Buba… https://bu… https:/… NA      Use our ap…           3          22
#>  9 bus_l… Bus … https://zl… NA       https:… Bus lanes …           6          33
#> 10 busy_… Busy… https://ju… https:/… https:… Easy editi…          11           2
#> # ℹ 220 more rows
#> # ℹ 2 more variables: unique_keys <int>, unique_tags <int>

# get all highway keys and tags that are used in a project
projects_keys("highway")
#> # A tibble: 9 × 5
#>   key                   projects in_wiki count_all count_all_fraction
#>   <chr>                    <int>   <int>     <int>              <dbl>
#> 1 abandoned:highway            4       2     81958             0     
#> 2 area:highway                14      10    323203             0     
#> 3 cycle_highway                1       3       859             0     
#> 4 disused:highway              7       1     39944             0     
#> 5 highway                     83      29 267151198             0.0251
#> 6 highway:virtual              1       0        83             0     
#> 7 highway_authority_ref        5       1     52310             0     
#> 8 highway_ref                  4       0      1047             0     
#> 9 horse_scale_highway          1       0         0             0     
projects_tags("highway")
#> # A tibble: 205 × 6
#>    key               value         projects in_wiki count_all count_all_fraction
#>    <chr>             <chr>            <int>   <int>     <int>              <dbl>
#>  1 abandoned:highway footway              1       0      1158                  0
#>  2 abandoned:highway motorway             1       0       817                  0
#>  3 abandoned:highway motorway_link        1       0       630                  0
#>  4 abandoned:highway path                 2       0     12752                  0
#>  5 abandoned:highway primary              2       0      2675                  0
#>  6 abandoned:highway residential          2       0      2711                  0
#>  7 abandoned:highway secondary            2       0      1567                  0
#>  8 abandoned:highway service              2       0      3009                  0
#>  9 abandoned:highway tertiary             2       0      1128                  0
#> 10 abandoned:highway track                2       0     44778                  0
#> # ℹ 195 more rows

#