Free search

Search for tags, keys or relation roles.

search_key_value(query, ...)

search_keyword(query, ...)

search_role(query, ...)

search_value(query, ...)

Source

https://taginfo.openstreetmap.org/taginfo/apidoc

Arguments

query

Substring to search for. Used to search key/value combinations in case of search_key_value, keywords in wiki pages in case of search_keyword, relation roles in case of search_role, and values in case of search_value

Value

For search_key_value and search_keyword returns a dataframe containing the matched key-value combinations. For search_role returns a dataframe containing the matched relation type and role.

Examples

search_key_value("highway=residential")
#> # A tibble: 139 × 3
#>    key                           value                                 count_all
#>    <chr>                         <chr>                                     <int>
#>  1 highway:historical            residential                                   1
#>  2 highway                       residential;footway                           1
#>  3 highway                       raced:residential                             1
#>  4 source:highway_classification appears to be of residential use only         1
#>  5 his:1957-:highway             residential                                   1
#>  6 was:proposed:highway          residential                                   1
#>  7 highway_authority_ref         residential                                   1
#>  8 highway_authority_ref         residential road                              1
#>  9 his:1948-:highway             residential                                   1
#> 10 his:1672-:highway             residential                                   1
#> # ℹ 129 more rows
search_key_value("amenity=")
#> # A tibble: 13,627 × 3
#>    key                    value                                        count_all
#>    <chr>                  <chr>                                            <int>
#>  1 fictional:amenity      marketplace                                          1
#>  2 fictional:amenity      cafe                                                 1
#>  3 fictional:amenity      police                                               1
#>  4 fictional:amenity      restaurant                                           1
#>  5 amenity:--2020-08-31   cafe                                                 1
#>  6 amenity:1995-2024      cafe                                                 1
#>  7 source:disused:amenity survey in 03/2016                                    1
#>  8 source:disused:amenity survey                                               1
#>  9 source:disused:amenity Another mapper's changeset comment; my reco…         1
#> 10 source:disused:amenity https://twitter.com/BirdBrasserie/status/14…         1
#> # ℹ 13,617 more rows
search_value("residential")
#> # A tibble: 11,820 × 3
#>    key              value                              count_all
#>    <chr>            <chr>                                  <int>
#>  1 surveillance     residential                                1
#>  2 development      residential                                1
#>  3 traffic_signals  signallanduse=residential                  1
#>  4 impasse          residential                                1
#>  5 addr:subdivision Sun Valley Residential Estates             1
#>  6 heritage         residential                                1
#>  7 int_name         Good Friends 3 Residential                 1
#>  8 int_name         Este Home Residential Development          1
#>  9 int_name         residential compound Green Village         1
#> 10 int_name         Kaoha Nui Residential apartments           1
#> # ℹ 11,810 more rows

search_keyword("fire")
#> # A tibble: 108 × 2
#>    key        value           
#>    <chr>      <chr>           
#>  1 actuator   "manual"        
#>  2 amenity    "fire_station"  
#>  3 amenity    "fuel"          
#>  4 amenity    "library"       
#>  5 amenity    "recycling"     
#>  6 amenity    "training"      
#>  7 barrier    "berm"          
#>  8 barrier    "cycle_barrier" 
#>  9 barrier    "retaining_wall"
#> 10 board_type ""              
#> # ℹ 98 more rows
search_keyword("railway")
#> # A tibble: 52 × 2
#>    key             value          
#>    <chr>           <chr>          
#>  1 amenity         "workshop"     
#>  2 bridge          "cantilever"   
#>  3 bridge:movable  "swing"        
#>  4 building        "train_station"
#>  5 club            ""             
#>  6 club            "model_railway"
#>  7 construction    ""             
#>  8 electrified     "4th_rail"     
#>  9 industrial      "logistics"    
#> 10 line_attachment "pulley"       
#> # ℹ 42 more rows

search_role("highway")
#> # A tibble: 13 × 3
#>    rtype            role                                               count_all
#>    <chr>            <chr>                                                  <int>
#>  1 site             Highway                                                    1
#>  2 public_transport area:highway                                               1
#>  3 route            highway service                                            1
#>  4 route            state highway                                              1
#>  5 route            highway_service                                            1
#>  6 route            access=private highway=service oneway=no service=…         1
#>  7 route            highway=path                                               1
#>  8 bridge           highway                                                    2
#>  9 route            Mumbai-Pune Expressway | Bangalore Mumbai Highway          9
#> 10 route            Texas State Highway Loop 332                              15
#> 11 area             highway                                                   20
#> 12 site             highway                                                   45
#> 13 level            highway                                                  480