OSM wiki

Retrieve data from the OSM wiki including languages and key lists. Additionally, ietf_languages returns a list of language tags from the IETF BCP47 registry.

wiki_data(query = NULL, lang = "en", ...)

wiki_languages()

ietf_languages(query = NULL, ...)

Arguments

query

A search term to filter the results by. Searches language names in case of ietf_languages and keys, values, relation types, item codes and item descriptions in case of wiki_data.

lang

Language of the description text. Defaults to english.

...

Additional generic parameters used for pagination, filtering, and sorting. Can include the following:

page: If pagination is enabled, specifies the page number.

rp: Enables pagination and specifies the number of results returned per page.

filter: A character vector specifying the OSM geometry type to filter the results by. For example, if "nodes", only tags on nodes are returned. Can be several of "all" (no filtering), "nodes", "relations", or "ways".

sortname: Numeric column of the output by which to arrange the results.

sortorder: Whether to sort ascending ("asc") or descending ("desc").

Note that not all additional arguments are defined for all API methods. For details, refer to api_spec.

Value

wiki_data returns a dataframe containing the keys and their descriptions, relation types, and values. wiki_languages returns a dataframe containing the native and english names as well as the number of key pages and the number of tag pages. ietf_languages returns a dataframe containing the defined languages as well as their type, description, time added and additional notes.

Examples

# Query languages
wiki_languages()
#> # A tibble: 75 × 8
#>    code  dir   native_name    english_name wiki_key_pages wiki_key_pages_fract…¹
#>    <chr> <chr> <chr>          <chr>                 <int>                  <dbl>
#>  1 af    ltr   Afrikaans      Afrikaans                 0               0       
#>  2 ar    auto  العربية        Arabic                    5               0.000875
#>  3 az    ltr   Azərbaycan di… Azerbaijani               0               0       
#>  4 bg    ltr   Български език Bulgarian                 4               0.000700
#>  5 bn    auto  বাংলা          Bengali                   0               0       
#>  6 br    auto  NA             NA                        1               0.000175
#>  7 bs    auto  NA             NA                        0               0       
#>  8 ca    ltr   Català         Catalan                  21               0.00367 
#>  9 cnr   auto  NA             NA                        1               0.000175
#> 10 cs    ltr   Čeština        Czech                   818               0.143   
#> # ℹ 65 more rows
#> # ℹ abbreviated name: ¹​wiki_key_pages_fraction
#> # ℹ 2 more variables: wiki_tag_pages <int>, wiki_tag_pages_fraction <dbl>
ietf_languages("irish")
#> # A tibble: 5 × 5
#>   type     subtag description             added      notes                 
#>   <chr>    <chr>  <chr>                   <chr>      <chr>                 
#> 1 Language ga     Irish                   2005-10-16 "Default script: Latn"
#> 2 Language isg    Irish Sign Language     2009-07-29 ""                    
#> 3 Language mga    Middle Irish (900-1200) 2005-10-16 ""                    
#> 4 Language pgl    Primitive Irish         2011-08-16 ""                    
#> 5 Language sga    Old Irish (to 900)      2005-10-16 ""                    

# Get all amenities defined on the wiki
wiki_data("amenity")
#> Error in wiki_data("amenity"): could not find function "wiki_data"